refac: migrate from react context to zustand store
