+++ backend/.env.example
@@ -74,3 +74,7 @@ REDIS_PASSWORD=
 # Sentry Configuration
 SENTRY_DSN=
 SENTRY_ENVIRONMENT=development
+
+
+
+SECURITY_ENCRYPTION_KEY=YOUR_SECURITY_ENCRYPTION_KEY
