+++ frontend/package.json
@@ -21,7 +21,8 @@
     "react-hook-form": "^7.71.2",
     "react-icons": "^5.5.0",
     "recharts": "^3.7.0",
-    "zod": "^4.3.6"
+    "zod": "^4.3.6",
+    "zustand": "^5.0.11"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4",
